Remove old comments and normalise style in esp32`s xively_if

+++ examples/esp32/xively_demo/main/xively_if.h
@@ -9,6 +9,7 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
 #include "xively.h"
 
 #define XIF_MQTT_TOPIC_MAX_LEN 256
@@ -18,19 +19,21 @@ typedef struct
     char button_topic[XIF_MQTT_TOPIC_MAX_LEN];
     char led_topic[XIF_MQTT_TOPIC_MAX_LEN];
 } xif_mqtt_topics_t;
+
 extern xif_mqtt_topics_t xif_mqtt_topics; /* Filled by xif_set_device_info */
 
-typedef enum /* Flags to request changes state machine actions - Ordered by priority */
+/* Actions that can be requested of the XIF state machine - Ordered by priority */
+typedef enum
 {
-    XIF_REQUEST_CONTINUE   = ( 1 << 0 ),
-    XIF_REQUEST_PAUSE      = ( 1 << 1 ),
-    XIF_REQUEST_SHUTDOWN   = ( 1 << 2 ),
-    XIF_REQUEST_ALL        = 0xff
+    XIF_REQUEST_CONTINUE = ( 1 << 0 ), /* Keep ticking libxively */
+    XIF_REQUEST_PAUSE    = ( 1 << 1 ), /* Stop ticking, wait for _CONTINUE */
+    XIF_REQUEST_SHUTDOWN = ( 1 << 2 ), /* Stop ticking, shutdown libxi and RTOS task */
+    XIF_REQUEST_ALL      = 0xff
 } xif_action_requests_t;
 
 /* Configure Xively credentials and topics
  */
-int xif_set_device_info( char* xi_acc_id, char* xi_dev_id, char* xi_dev_pwd );
+int8_t xif_set_device_info( char* xi_acc_id, char* xi_dev_id, char* xi_dev_pwd );
 
 /* Xively Interface event loop - It handles the MQTT library's events loop and
  * coordinates the actions requested/kickstarted from other RTOS tasks. Loops
@@ -42,9 +45,9 @@ void xif_rtos_task( void* param );
  * XIF will be paused after the disconnection - Call xif_events_continue to
  * unpause the events loop and re-connect to the broker
  */
-int xif_disconnect( void );
+int8_t xif_disconnect( void );
 
-/* 
+/*
  * xif_connect() is called internally from xif_rtos_task before entering the
  * events loop, so this function doesn't need to be called when starting the
  * Xively Interface
@@ -54,22 +57,28 @@ int xif_disconnect( void );
  * The application only needs this function to reconnect after calling
  * xif_disconnect()
  *
- * @retval: -1: xi_connect failed
- *           0: 0 xi_connect OK
- *           1: xi_connect failed because the connection was already initialized
+ * @retval -1: xi_connect failed
+ * @retval  0: 0 xi_connect OK
+ * @retval  1: xi_connect failed because the connection was already initialized
  */
-int xif_connect( void );
+int8_t xif_connect( void );
 
 void xif_publish_button_state( int input_level );
 
-int xif_request_action( xif_action_requests_t requested_action );
+/* Request an action from the Xively Interface's state machine.
+ * xif_action_requests_t is ordered by priority (highest value, highes priority)
+ *
+ * @retval -1: Error
+ * @retval  0: OK
+ */
+int8_t xif_request_action( xif_action_requests_t requested_action );
 
 /* Sample implementation declared WEAK in xively_if.c so you can overwrite it */
 extern void xif_recv_mqtt_msg_callback( const xi_sub_call_params_t* const params );
 
 /* Query the MQTT connection status (as far as the TCP/MQTT layers are aware)
  */
-int xif_is_connected( void );
+int8_t xif_is_connected( void );
 
 /* This callback is __weak__ in xively_if.c so you can overwrite it with your own.
  * For this demo, we permanently shut down the Xively Interface when we get
